Add ability to edit notes with external editor
If -z or --append is passed, then the external
editor is opened. The written text is appended
to the current note using the delimiter.

If no arguments are passed, then the current
note is opened in the editor for editing and
the entire note is overwritten.

There is one small issue, which is that the
printing of notes is a little ugly because
of the line breaks.. I'm not quite sure how
simple that is going to be to fix.